pyproject.toml
[tool.poetry]
name = "goon-files"
version = "0.1.0"
description = "Auth database service for Goon Auth Network"
authors = ["Ian P <ian@aotc.us>"]
repository = "https://github.com/GoonAuthNetwork/goon-files"
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "1.8.1"
fastapi = "0.63.0"
uvicorn = "0.13.4"
gunicorn = "20.0.4"
motor = "^2.3.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
